Use RPATH for the unit tests only
This allows the unit tests to run without having to specific the
location of the LLVM libraries using LD_LIBRARY_PATH.
It does not matter that the path in embedded in the RPATH since unit
tests are never installed.
This is purely a convenience to execute without much fuss in the build
tree.